requirements.txt
atlassian-python-api
pyyaml 